create routes.js for decouple routes from server.js file
